# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Influencer Butler Public API",
    "version": "0.1.0",
    "description": "Public HTTP endpoints exposed by influencerbutler.com. Authenticated routes use Supabase session cookies; programmatic agent access uses bearer tokens issued from /dashboard/api-keys.",
    "contact": {
      "name": "Influencer Butler",
      "url": "https://www.influencerbutler.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://www.influencerbutler.com/legal/terms"
    }
  },
  "servers": [
    { "url": "https://www.influencerbutler.com", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "supabaseSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie set after login."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ib_pat_<random>",
        "description": "Personal access token issued at /dashboard/api-keys."
      }
    },
...
  },
  "paths": {
...
    "/api/checkout/guest": {
      "post": {
        "tags": ["billing"],
        "summary": "Begin a Lemon Squeezy guest checkout session",
        "responses": {
          "200": { "description": "Returns a hosted checkout URL." },
          "400": { "description": "Invalid checkout payload." }
        }
      }
    },
...
    "/api/mcp": {
      "post": {
        "tags": ["agents"],
        "summary": "Streamable HTTP MCP transport endpoint",
        "description": "Implements the Model Context Protocol streamable-http transport. Tool catalog is published at /.well-known/mcp.json.",
        "security": [{ "bearerToken": [] }],
        "responses": {
          "200": { "description": "MCP response stream." },
          "401": { "description": "Missing or invalid bearer token." }
        }
      }
    },
    "/api/skills/{id}/invoke": {
      "post": {
        "tags": ["agents"],
        "summary": "Invoke an agent skill by id",
        "description": "Skill catalog is published at /.well-known/agent-skills.json. Anonymous skills require no auth; user-scoped skills require a bearer token.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Skill id (e.g. feature-catalog, pricing-tiers, amazon-affiliate-link-creation, earnings-summary)."
          }
        ],
        "security": [{}, { "bearerToken": [] }],
        "responses": {
          "200": { "description": "Skill output." },
          "400": { "description": "Invalid input." },
          "401": { "description": "Bearer token required for this skill." },
          "404": { "description": "Unknown skill id." }
        }
      }
    },
    "/api/mcp-keys": {
      "get": {
        "tags": ["agents"],
        "summary": "List the authenticated user's personal access tokens",
        "security": [{ "supabaseSession": [] }],
        "responses": {
          "200": { "description": "Array of token metadata (no secret values)." },
          "401": { "description": "Not authenticated." }
        }
      },
      "post": {
        "tags": ["agents"],
        "summary": "Issue a new personal access token",
        "security": [{ "supabaseSession": [] }],
        "responses": {
          "200": { "description": "Token returned once; record it immediately." },
          "401": { "description": "Not authenticated." }
        }
      }
    }
  }
}